% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulo.R
\name{modulo}
\alias{modulo}
\title{modulo de dos numeros}
\usage{
modulo(a, b)
}
\arguments{
\item{a}{un elemento numerico}

\item{b}{un elemento numerico distinto de 0}
}
\description{
Esta es una funcion de ejemplo para calcular \code{a} modulo \code{b}.
}
\details{
La funcion recibe dos elementos \code{a}, \code{b} que son un numero y devuelve
su modulo. La documentacion de ayuda es generada usando roxygen2.
}
\examples{
modulo(3,2)
}
