% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplicacion.R
\name{multiplicacion}
\alias{multiplicacion}
\title{multiplicación de dos numeros}
\usage{
multiplicacion(a, b)
}
\arguments{
\item{a}{un elemento o vector numerico}

\item{b}{un elemento o vector numerico}
}
\description{
Esta es una funcion de ejemplo para multiplicar dos valores o vectores \code{a}, \code{b}.
}
\details{
La funcion recibe dos elementos \code{a}, \code{b} que pueden ser un numero o un vector numerico y devuelve
su multiplicación. La documentacion de ayuda es generada usando roxygen2.
}
\examples{
multiplicacion(3,2)
multiplicacion(0:10,0:10)
}
