% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suma.R
\name{suma}
\alias{suma}
\title{Suma de dos numeros}
\usage{
suma(a, b)
}
\arguments{
\item{a}{un elemento o vector numerico}

\item{b}{un elemento o vector numerico}
}
\description{
Esta es una funcion de ejemplo para sumar dos valores o vectores \code{a}, \code{b}.
}
\details{
La funcion recibe dos elementos \code{a}, \code{b} que pueden ser un numero o un vector numerico y devuelve
su suma. La documentacion de ayuda es generada usando roxygen2.
}
\examples{
suma(3,2)
suma(0:10,0:10)
}
